Refactor CLI commands and add success summaries
