+++ src/cli/cli.ts
@@ -1,26 +1,17 @@
-import { access, readdir } from 'node:fs/promises';
-import { basename, join, relative } from 'node:path';
+/* eslint-disable max-lines */
+import { readdir } from 'node:fs/promises';
+import { join } from 'node:path';
 
 import { cac } from 'cac';
 
 import pkg from '../../package.json';
-import { collectClarifications } from '../agent/clarify';
 import { createSessionPool } from '../agent/session';
 import {
   type AssistSuggestion,
   suggestCliRecovery,
   suggestConfigRecovery,
 } from '../ai/cognition/assist';
-import {
-  createConversationStore,
-  exportConversationSnapshot,
-  loadConversationSnapshot,
-  renderConversationSummary,
-  summarizeConversationSnapshot,
-} from '../ai/conversation';
-import { benchmarkCognitionModels } from '../ai/model-benchmark';
-import { applyCognitionModelRouting, buildModelRoutingReport } from '../ai/model-routing';
-import { buildAnalyticsReport } from '../analytics/analytics';
+import { applyCognitionModelRouting } from '../ai/model-routing';
 import { type EnvLoadSummary, getLoadedEnvSummary } from '../config/env';
 import {
   applyInitSuggestions,
@@ -33,54 +24,16 @@ import {
 } from '../config/init';
 import { loadConfig } from '../config/load';
 import type { Config, ConfigInput } from '../config/schema';
-import { requireGitHubAuth, requireGitHubConfig } from '../config/validate';
 import type { RunContext } from '../core/context';
 import { withRunContext } from '../core/context';
 import { normalizeError, SilvanError } from '../core/errors';
 import { detectRepoContext } from '../core/repo';
 import { formatRepoLabel } from '../core/repo-label';
-import {
-  resumeRun,
-  runImplementation,
-  runLearningNotes,
-  runPlanner,
-  runRecovery,
-  runReviewLoop,
-} from '../core/run-controller';
+import { runPlanner } from '../core/run-controller';
 import { createEnvelope } from '../events/emit';
-import type { Event, EventMode, RunStep } from '../events/schema';
-import { runGit } from '../git/exec';
-import {
-  createWorktree,
-  hasUncommittedChanges,
-  installDependencies,
-  listWorktrees,
-  lockWorktree,
-  normalizeClaudeSettings,
-  pruneWorktrees,
-  rebaseOntoBase,
-  removeWorktree,
-  unlockWorktree,
-} from '../git/worktree';
-import { findMergedPr } from '../github/pr';
-import { findHelpTopic, listHelpTopics } from '../help/topics';
-import { evaluateLearningTargets } from '../learning/auto-apply';
-import { applyLearningNotes } from '../learning/notes';
+import type { EventMode, RunStep } from '../events/schema';
 import { PRIORITY_MAX, PRIORITY_MIN } from '../queue/priority';
-import {
-  deriveConvergenceFromSnapshot,
-  loadRunSnapshot,
-  markRunAborted,
-  writeOverrideArtifact,
-} from '../run/controls';
-import type { ArtifactEntry } from '../state/artifacts';
-import { readArtifact } from '../state/artifacts';
-import {
-  type LearningRequest,
-  listLearningRequests,
-  readLearningRequest,
-  writeLearningRequest,
-} from '../state/learning';
+import { deriveConvergenceFromSnapshot } from '../run/controls';
 import {
   loadOnboardingState,
   markFirstRunCompleted,
@@ -90,24 +43,24 @@ import { initStateStore } from '../state/store';
 import { type LocalTaskInput } from '../task/providers/local';
 import { resolveTask } from '../task/resolve';
 import { confirmAction } from '../utils/confirm';
-import { sanitizeName } from '../utils/slug';
-import { parseTimeInput } from '../utils/time';
-import { buildAnalyticsNextSteps, renderAnalyticsReport } from './analytics-output';
-import { parseAnswerPairs } from './answers';
+import { registerAgentCommands } from './commands/agent';
+import { registerAnalyticsCommands } from './commands/analytics';
+import { registerCompletionCommand } from './commands/completion';
 import { registerConfigCommands } from './commands/config';
+import { registerConversationCommands } from './commands/conversation';
 import { registerDoctorCommands } from './commands/doctor';
+import { registerHelpCommand } from './commands/help';
+import { registerLearningCommands } from './commands/learning';
+import { registerLogCommands } from './commands/logs';
+import { registerModelCommands } from './commands/models';
 import { registerQueueCommands } from './commands/queue';
 import { registerReviewCommands } from './commands/review';
+import { registerRunCommands } from './commands/run';
 import { registerTaskCommands, startTaskFlow } from './commands/task';
+import { registerTreeCommands } from './commands/tree';
 import { registerUiCommands } from './commands/ui';
-import {
-  generateBashCompletion,
-  generateFishCompletion,
-  generateZshCompletion,
-} from './completion';
 import { renderCliError } from './errors';
 import { buildHelpSections } from './help-output';
-import { renderHelpTopic, renderHelpTopicsList } from './help-topics-output';
 import {
   renderInitAssist,
   renderInitDetection,
@@ -117,17 +70,7 @@ import {
 } from './init-output';
 import { emitJsonError, emitJsonResult, emitJsonSuccess } from './json-output';
 import { buildEmitContext, createCliLogger } from './logger';
-import {
-  renderModelBenchmarkReport,
-  renderModelRoutingReport,
-} from './model-routing-output';
-import {
-  formatKeyList,
-  formatKeyValues,
-  formatStatusLabel,
-  renderSectionHeader,
-  renderSuccessSummary,
-} from './output';
+import { renderSectionHeader } from './output';
 import {
   type QuickstartCheck,
   type QuickstartRunSummary,
@@ -139,11 +82,7 @@ import {
   renderReturningSummary,
   renderWorkflowOverview,
 } from './quickstart-output';
-import {
-  renderRunListMinimal,
-  renderRunListTable,
-  type RunListEntry,
-} from './run-list-output';
+import { deriveRunListStatus } from './run-list-output';
 import {
   renderNextSteps,
   renderPlanSummary,
@@ -152,7 +91,6 @@ import {
   summarizePlan,
 } from './task-start-output';
 import type { CliOptions } from './types';
-import { deriveWorktreeStatus, renderWorktreeListTable } from './worktree-list-output';
 
 const cli = cac('silvan');
 
@@ -232,6 +170,7 @@ cli
   .command('init', 'Initialize silvan.config.ts with guided prompts')
   .option('--yes', 'Skip prompts and use defaults')
   .option('--assist', 'Use cognition to suggest defaults')
+  // eslint-disable-next-line complexity
   .action(async (options: CliOptions) => {
     const configResult = await loadConfig(buildConfigOverrides(options), {
       cwd: process.cwd(),
@@ -337,6 +276,7 @@ cli
   .command('quickstart', 'Guided setup and sample plan')
   .option('--yes', 'Skip prompts and use defaults')
   .option('--assist', 'Use cognition to suggest defaults')
+  // eslint-disable-next-line complexity
   .action(async (options: CliOptions) => {
     const jsonMode = Boolean(options.json);
     const showOutput = !options.quiet && !jsonMode;
@@ -640,19 +580,19 @@ cli
     }
   });
 
-function parseNumberFlag(value: string | undefined): number | undefined {
-  if (!value) return undefined;
+function parseNumberFlag(value: string | undefined): number | null {
+  if (!value) return null;
   const parsed = Number(value);
-  return Number.isFinite(parsed) ? parsed : undefined;
+  return Number.isFinite(parsed) ? parsed : null;
 }
 
-function parseListFlag(value: string | undefined): string[] | undefined {
-  if (!value) return undefined;
+function parseListFlag(value: string | undefined): string[] | null {
+  if (!value) return null;
   const items = value
     .split(',')
     .map((item) => item.trim().toLowerCase())
     .filter((item) => item.length > 0);
-  return items.length > 0 ? items : undefined;
+  return items.length > 0 ? items : null;
 }
 
 function parseCsvFlag(value: string | undefined): string[] | undefined {
@@ -664,13 +604,13 @@ function parseCsvFlag(value: string | undefined): string[] | undefined {
   return items.length > 0 ? items : undefined;
 }
 
-function parseModelList(value: string | undefined): string[] | undefined {
-  if (!value) return undefined;
+function parseModelList(value: string | undefined): string[] | null {
+  if (!value) return null;
   const items = value
     .split(',')
     .map((item) => item.trim())
     .filter((item) => item.length > 0);
-  return items.length > 0 ? items : undefined;
+  return items.length > 0 ? items : null;
 }
 
 function collectBenchmarkModels(config: Config): string[] {
@@ -726,21 +666,6 @@ function parseQueuePriority(value: string | undefined, fallback: number): number
     });
   }
   return parsed;
-}
-
-function deriveRunListStatus(
-  runStatus: string | undefined,
-  convergenceStatus: string,
-): string {
-  if (runStatus === 'success') return 'success';
-  if (runStatus === 'failed') return 'failed';
-  if (runStatus === 'canceled') return 'canceled';
-  if (convergenceStatus === 'blocked') return 'blocked';
-  if (convergenceStatus.startsWith('waiting_')) return 'blocked';
-  if (convergenceStatus === 'converged') return 'success';
-  if (convergenceStatus === 'failed') return 'failed';
-  if (convergenceStatus === 'aborted') return 'canceled';
-  return runStatus ?? 'running';
 }
 
 type CognitionKey = {
@@ -895,6 +820,7 @@ async function maybeSuggestCliRecovery(options: {
   }
 }
 
+// eslint-disable-next-line complexity
 function buildConfigOverrides(options: CliOptions): ConfigInput {
   const overrides: ConfigInput = {};
 
@@ -1115,530 +1041,15 @@ async function withCliContext<T>(
   );
 }
 
-cli
-  .command('help [topic]', 'View help topics and concepts')
-  .action(async (topic: string | undefined, options: CliOptions) => {
-    const topics = listHelpTopics();
-    const jsonMode = Boolean(options.json);
+registerHelpCommand(cli);
 
-    if (!topic) {
-      if (jsonMode) {
-        await emitJsonSuccess({
-          command: 'help',
-          data: {
-            topics: topics.map(({ id, title, summary, category }) => ({
-              id,
-              title,
-              summary,
-              category,
-            })),
-            usage: 'silvan help <topic>',
-            commandHelp: 'silvan <command> --help',
-          },
-        });
-        return;
-      }
-
-      if (options.quiet) {
-        return;
-      }
-
-      console.log(renderHelpTopicsList(topics));
-      return;
-    }
-
-    const matched = findHelpTopic(topic);
-    if (!matched) {
-      throw new SilvanError({
-        code: 'help.topic_not_found',
-        message: `Unknown help topic: ${topic}`,
-        userMessage: `Unknown help topic: ${topic}`,
-        kind: 'validation',
-        nextSteps: [
-          'Run `silvan help` to list available topics.',
-          'Run `silvan --help` for command help.',
-        ],
-      });
-    }
-
-    if (jsonMode) {
-      await emitJsonSuccess({
-        command: 'help',
-        data: {
-          topic: {
-            id: matched.id,
-            title: matched.title,
-            summary: matched.summary,
-            category: matched.category,
-            intro: matched.intro,
-            sections: matched.sections,
-            examples: matched.examples ?? [],
-            seeAlso: matched.seeAlso ?? [],
-          },
-        },
-      });
-      return;
-    }
-
-    if (options.quiet) {
-      return;
-    }
-
-    console.log(renderHelpTopic(matched));
-  });
+registerHelpCommand(cli);
 
 // Worktree commands - aliased as both 'tree' and 't'
-cli.command('tree list', 'List all git worktrees').action(async (options: CliOptions) => {
-  const mode: EventMode = options.json ? 'json' : 'headless';
-  await withCliContext(options, mode, async (ctx) => {
-    const logger = createCliLogger(ctx);
-    const worktrees = await runStep(
-      ctx,
-      'git.worktree.list',
-      'List worktrees',
-      async () =>
-        listWorktrees({
-          repoRoot: ctx.repo.repoRoot,
-          bus: ctx.events.bus,
-          context: { runId: ctx.runId, repoRoot: ctx.repo.repoRoot, mode },
-          includeStatus: true,
-        }),
-    );
-
-    const nextSteps = ['silvan tree add <name>', 'silvan tree remove <name>'];
-
-    if (options.json) {
-      await emitJsonSuccess({
-        command: 'tree list',
-        data: {
-          total: worktrees.length,
-          worktrees: worktrees.map((worktree) => ({
-            ...worktree,
-            status: deriveWorktreeStatus(worktree),
-          })),
-        },
-        nextSteps,
-        runId: ctx.runId,
-        repoRoot: ctx.repo.repoRoot,
-      });
-      return;
-    }
-
-    if (options.quiet) {
-      return;
-    }
-
-    const lines: string[] = [];
-    lines.push(renderWorktreeListTable(worktrees, { total: worktrees.length }));
-    lines.push(renderNextSteps(nextSteps));
-    await logger.info(lines.join('\n'));
-  });
+registerTreeCommands(cli, {
+  withCliContext,
+  runStep,
 });
-
-cli
-  .command('tree add <name>', 'Create a new worktree with branch')
-  .action(async (name: string, options: CliOptions) => {
-    const mode: EventMode = options.json ? 'json' : 'headless';
-    await withCliContext(options, mode, async (ctx) => {
-      const logger = createCliLogger(ctx);
-      const safeName = sanitizeName(name);
-
-      const worktree = await runStep(
-        ctx,
-        'git.worktree.create',
-        'Create worktree',
-        async () =>
-          createWorktree({
-            repoRoot: ctx.repo.repoRoot,
-            name: safeName,
-            config: ctx.config,
-            bus: ctx.events.bus,
-            context: { runId: ctx.runId, repoRoot: ctx.repo.repoRoot, mode },
-          }),
-      );
-
-      await normalizeClaudeSettings({ worktreePath: worktree.path });
-
-      const installResult = await runStep(
-        ctx,
-        'deps.install',
-        'Installing dependencies',
-        () => installDependencies({ worktreePath: worktree.path }),
-      );
-      if (!installResult.ok) {
-        await logger.warn(`Warning: bun install failed in ${worktree.path}`, {
-          stderr: installResult.stderr,
-          stdout: installResult.stdout,
-        });
-      }
-
-      await logger.info(
-        renderSuccessSummary({
-          title: `Created worktree '${safeName}'`,
-          details: [
-            ['Path', worktree.path],
-            ['Branch', worktree.branch ?? safeName],
-          ],
-          nextSteps: [`cd ${worktree.path}`, 'silvan task start "Your task"'],
-        }),
-      );
-    });
-  });
-
-cli
-  .command('tree remove [name]', 'Remove a worktree')
-  .option('--force', 'Force removal even if dirty')
-  .option('--task <task>', 'Remove worktree for a task reference')
-  .action(async (name: string | undefined, options: CliOptions) => {
-    const mode: EventMode = options.json ? 'json' : 'headless';
-    await withCliContext(options, mode, async (ctx) => {
-      const logger = createCliLogger(ctx);
-      const task = options.task ? sanitizeName(options.task) : undefined;
-      const targetName = name ? sanitizeName(name) : undefined;
-      if (!task && !targetName) {
-        throw new Error('Worktree name or --task is required.');
-      }
-
-      const worktrees = await runStep(
-        ctx,
-        'git.worktree.list',
-        'List worktrees',
-        async () =>
-          listWorktrees({
-            repoRoot: ctx.repo.repoRoot,
-            bus: ctx.events.bus,
-            context: { runId: ctx.runId, repoRoot: ctx.repo.repoRoot, mode },
-            includeStatus: true,
-          }),
-      );
-
-      const expectedPath = task
-        ? join(ctx.repo.repoRoot, ctx.config.naming.worktreeDir, task)
-        : null;
-      const expectedBranch = task ? `${ctx.config.naming.branchPrefix}${task}` : null;
-      const targets = worktrees.filter((worktree) => {
-        if (expectedPath && worktree.path === expectedPath) return true;
-        if (expectedBranch && worktree.branch === expectedBranch) return true;
-        if (targetName && worktree.branch === targetName) return true;
-        if (targetName && basename(worktree.path) === targetName) return true;
-        return false;
-      });
-
-      if (targets.length === 0) {
-        throw new Error(`Worktree not found: ${task ?? targetName}`);
-      }
-      if (targets.length > 1) {
-        const paths = targets.map((target) => target.path).join(', ');
-        throw new Error(`Worktree name is ambiguous: ${task ?? targetName} (${paths})`);
-      }
-
-      const target = targets[0]!;
-
-      if (target.isDirty && !options.force) {
-        throw new Error(`Worktree ${name} is dirty. Use --force to remove.`);
-      }
-
-      if (!options.yes) {
-        const confirmed = await confirmAction(`Remove worktree ${target.path}?`);
-        if (!confirmed) {
-          await logger.info('Canceled.');
-          return;
-        }
-      }
-
-      await runStep(ctx, 'git.worktree.remove', 'Remove worktree', async () =>
-        removeWorktree({
-          repoRoot: ctx.repo.repoRoot,
-          path: target.path,
-          force: Boolean(options.force),
-          bus: ctx.events.bus,
-          context: { runId: ctx.runId, repoRoot: ctx.repo.repoRoot, mode },
-        }),
-      );
-
-      await logger.info(
-        renderSuccessSummary({
-          title: 'Removed worktree',
-          details: [
-            ['Path', target.path],
-            ['Branch', target.branch ?? 'unknown'],
-          ],
-          nextSteps: ['silvan tree list', 'silvan tree add <name>'],
-        }),
-      );
-    });
-  });
-
-cli
-  .command('tree clean', 'Remove worktrees with merged PRs')
-  .option('--force', 'Force removal even if dirty')
-  .option('--all', 'Remove all merged worktrees without prompting')
-  .action(async (options: CliOptions) => {
-    const mode: EventMode = options.json ? 'json' : 'headless';
-    await withCliContext(options, mode, async (ctx) => {
-      const logger = createCliLogger(ctx);
-      const githubToken = requireGitHubAuth(ctx.config);
-      const github = await requireGitHubConfig({
-        config: ctx.config,
-        repoRoot: ctx.repo.repoRoot,
-        bus: ctx.events.bus,
-        context: { runId: ctx.runId, repoRoot: ctx.repo.repoRoot, mode },
-      });
-
-      const worktrees = await runStep(
-        ctx,
-        'git.worktree.list',
-        'List worktrees',
-        async () =>
-          listWorktrees({
-            repoRoot: ctx.repo.repoRoot,
-            bus: ctx.events.bus,
-            context: { runId: ctx.runId, repoRoot: ctx.repo.repoRoot, mode },
-          }),
-      );
-
-      const defaultBranch = ctx.config.repo.defaultBranch;
-      const candidates = worktrees.filter((worktree) => {
-        if (worktree.path === ctx.repo.gitRoot) return false;
-        if (!worktree.branch || worktree.branch === '(detached)') return false;
-        if (worktree.branch === defaultBranch) return false;
-        return true;
-      });
-
-      const merged: Array<{
-        worktree: (typeof candidates)[number];
-        pr: NonNullable<Awaited<ReturnType<typeof findMergedPr>>>;
-      }> = [];
-      for (const worktree of candidates) {
-        if (!worktree.branch) continue;
-        const pr = await findMergedPr({
-          owner: github.owner,
-          repo: github.repo,
-          headBranch: worktree.branch,
-          token: githubToken,
-          bus: ctx.events.bus,
-          context: { runId: ctx.runId, repoRoot: ctx.repo.repoRoot, mode },
-        });
-        if (pr) {
-          merged.push({ worktree, pr });
-        }
-      }
-
-      if (merged.length === 0) {
-        await logger.info(
-          renderSuccessSummary({
-            title: 'No merged worktrees found',
-            details: [['Candidates', `${candidates.length} worktree(s)`]],
-            nextSteps: ['silvan tree list'],
-          }),
-        );
-        return;
-      }
-
-      const removed: string[] = [];
-      const skipped: string[] = [];
-
-      for (const candidate of merged) {
-        const isDirty = await hasUncommittedChanges({
-          worktreePath: candidate.worktree.path,
-          bus: ctx.events.bus,
-          context: { runId: ctx.runId, repoRoot: ctx.repo.repoRoot, mode },
-        });
-        if (isDirty && !options.force) {
-          skipped.push(`${candidate.worktree.path} (dirty)`);
-          continue;
-        }
-
-        let shouldRemove = Boolean(options.yes || options.all);
-        if (!shouldRemove) {
-          shouldRemove = await confirmAction(
-            `Remove worktree ${candidate.worktree.path} (PR #${candidate.pr.number})?`,
-          );
-        }
-
-        if (!shouldRemove) {
-          skipped.push(`${candidate.worktree.path} (skipped)`);
-          continue;
-        }
-
-        await runStep(ctx, 'git.worktree.remove', 'Remove worktree', async () =>
-          removeWorktree({
-            repoRoot: ctx.repo.repoRoot,
-            path: candidate.worktree.path,
-            force: Boolean(options.force),
-            bus: ctx.events.bus,
-            context: { runId: ctx.runId, repoRoot: ctx.repo.repoRoot, mode },
-          }),
-        );
-        removed.push(candidate.worktree.path);
-      }
-
-      const remaining = Math.max(0, worktrees.length - removed.length);
-      const summaryLines: string[] = [];
-      summaryLines.push(
-        renderSectionHeader('Cleaned worktrees', { width: 60, kind: 'minor' }),
-      );
-      summaryLines.push(
-        ...formatKeyValues(
-          [
-            ['Removed', `${removed.length} worktree(s)`],
-            ['Skipped', `${skipped.length} worktree(s)`],
-            ['Remaining', `${remaining} worktree(s)`],
-          ],
-          { labelWidth: 12 },
-        ),
-      );
-      if (removed.length > 0) {
-        summaryLines.push(
-          ...formatKeyList('Removed', `${removed.length} worktree(s)`, removed, {
-            labelWidth: 12,
-          }),
-        );
-      }
-      if (skipped.length > 0) {
-        summaryLines.push(
-          ...formatKeyList('Skipped', `${skipped.length} worktree(s)`, skipped, {
-            labelWidth: 12,
-          }),
-        );
-      }
-      summaryLines.push(renderNextSteps(['silvan tree list']));
-      await logger.info(summaryLines.join('\n'));
-    });
-  });
-
-cli
-  .command('tree prune', 'Prune stale worktree data')
-  .action(async (options: CliOptions) => {
-    const mode: EventMode = options.json ? 'json' : 'headless';
-    await withCliContext(options, mode, async (ctx) => {
-      const logger = createCliLogger(ctx);
-      await runStep(ctx, 'git.worktree.prune', 'Prune worktrees', async () =>
-        pruneWorktrees({
-          repoRoot: ctx.repo.repoRoot,
-          bus: ctx.events.bus,
-          context: { runId: ctx.runId, repoRoot: ctx.repo.repoRoot, mode },
-        }),
-      );
-      await logger.info(
-        renderSuccessSummary({
-          title: 'Pruned worktree data',
-          details: [['Repository', ctx.repo.repoRoot]],
-          nextSteps: ['silvan tree list'],
-        }),
-      );
-    });
-  });
-
-cli
-  .command('tree lock <name>', 'Lock a worktree')
-  .option('--reason <reason>', 'Reason for locking')
-  .action(async (name: string, options: CliOptions & { reason?: string }) => {
-    const mode: EventMode = options.json ? 'json' : 'headless';
-    await withCliContext(options, mode, async (ctx) => {
-      const logger = createCliLogger(ctx);
-      const worktrees = await listWorktrees({
-        repoRoot: ctx.repo.repoRoot,
-        bus: ctx.events.bus,
-        context: { runId: ctx.runId, repoRoot: ctx.repo.repoRoot, mode },
-      });
-      const target = worktrees.find(
-        (worktree) => worktree.branch === name || basename(worktree.path) === name,
-      );
-      if (!target) {
-        throw new Error(`Worktree not found: ${name}`);
-      }
-      await runStep(ctx, 'git.worktree.lock', 'Lock worktree', async () =>
-        lockWorktree({
-          repoRoot: ctx.repo.repoRoot,
-          path: target.path,
-          ...(options.reason ? { reason: options.reason } : {}),
-          bus: ctx.events.bus,
-          context: { runId: ctx.runId, repoRoot: ctx.repo.repoRoot, mode },
-        }),
-      );
-      await logger.info(
-        renderSuccessSummary({
-          title: 'Locked worktree',
-          details: [
-            ['Path', target.path],
-            ['Branch', target.branch ?? 'unknown'],
-          ],
-          nextSteps: ['silvan tree list', 'silvan tree unlock <name>'],
-        }),
-      );
-    });
-  });
-
-cli
-  .command('tree unlock <name>', 'Unlock a worktree')
-  .action(async (name: string, options: CliOptions) => {
-    const mode: EventMode = options.json ? 'json' : 'headless';
-    await withCliContext(options, mode, async (ctx) => {
-      const logger = createCliLogger(ctx);
-      const worktrees = await listWorktrees({
-        repoRoot: ctx.repo.repoRoot,
-        bus: ctx.events.bus,
-        context: { runId: ctx.runId, repoRoot: ctx.repo.repoRoot, mode },
-      });
-      const target = worktrees.find(
-        (worktree) => worktree.branch === name || basename(worktree.path) === name,
-      );
-      if (!target) {
-        throw new Error(`Worktree not found: ${name}`);
-      }
-      await runStep(ctx, 'git.worktree.unlock', 'Unlock worktree', async () =>
-        unlockWorktree({
-          repoRoot: ctx.repo.repoRoot,
-          path: target.path,
-          bus: ctx.events.bus,
-          context: { runId: ctx.runId, repoRoot: ctx.repo.repoRoot, mode },
-        }),
-      );
-      await logger.info(
-        renderSuccessSummary({
-          title: 'Unlocked worktree',
-          details: [
-            ['Path', target.path],
-            ['Branch', target.branch ?? 'unknown'],
-          ],
-          nextSteps: ['silvan tree list'],
-        }),
-      );
-    });
-  });
-
-cli
-  .command('tree rebase', 'Rebase current branch onto base')
-  .action(async (options: CliOptions) => {
-    const mode: EventMode = options.json ? 'json' : 'headless';
-    await withCliContext(options, mode, async (ctx) => {
-      const logger = createCliLogger(ctx);
-      const ok = await runStep(ctx, 'git.rebase', 'Rebase onto base', async () =>
-        rebaseOntoBase({
-          repoRoot: ctx.repo.repoRoot,
-          baseBranch: ctx.config.repo.defaultBranch,
-          bus: ctx.events.bus,
-          context: { runId: ctx.runId, repoRoot: ctx.repo.repoRoot, mode },
-        }),
-      );
-
-      if (!ok) {
-        throw new Error('Rebase failed; conflicts were aborted.');
-      }
-
-      await logger.info(
-        renderSuccessSummary({
-          title: 'Rebase complete',
-          details: [
-            ['Branch', ctx.repo.branch ?? 'current'],
-            ['Base', ctx.config.repo.defaultBranch],
-          ],
-          nextSteps: ['git status', 'silvan pr open'],
-        }),
-      );
-    });
-  });
 
 registerReviewCommands(cli, {
   withCliContext,
@@ -1646,1527 +1057,35 @@ registerReviewCommands(cli, {
   persistRunState,
 });
 
-cli
-  .command('run list', 'List all recorded runs')
-  .option('--format <format>', 'table, minimal, or json', { default: 'table' })
-  .option('--status <status>', 'Filter by status (comma-separated)')
-  .option('--phase <phase>', 'Filter by phase (comma-separated)')
-  .option('--source <source>', 'Filter by task source (comma-separated)')
-  .option('--limit <n>', 'Number of runs to show', { default: '20' })
-  .option('--offset <n>', 'Skip the first N runs', { default: '0' })
-  .option('--show-source', 'Include task source column')
-  .action(
-    async (
-      options: CliOptions & {
-        format?: string;
-        status?: string;
-        phase?: string;
-        source?: string;
-        limit?: string;
-        offset?: string;
-        showSource?: boolean;
-      },
-    ) => {
-      const configResult = await loadConfig(buildConfigOverrides(options), {
-        cwd: process.cwd(),
-      });
-      const repo = await detectRepoContext({ cwd: configResult.projectRoot });
-      const state = await initStateStore(repo.projectRoot, {
-        lock: false,
-        mode: configResult.config.state.mode,
-        ...(configResult.config.state.root
-          ? { root: configResult.config.state.root }
-          : {}),
-        metadataRepoRoot: repo.gitRoot,
-      });
-      const runEntries = await readdir(state.runsDir);
-      const files = runEntries.filter((entry) => entry.endsWith('.json'));
-      const runs: RunListEntry[] = [];
-
-      for (const file of files) {
-        const runId = file.replace(/\.json$/, '');
-        const snapshot = await state.readRunState(runId);
-        if (!snapshot) continue;
-        const data = snapshot.data as Record<string, unknown>;
-        const run = (
-          typeof data['run'] === 'object' && data['run'] ? data['run'] : {}
-        ) as {
-          status?: string;
-          phase?: string;
-          step?: string;
-          updatedAt?: string;
-          startedAt?: string;
-        };
-        const summary = (
-          typeof data['summary'] === 'object' && data['summary'] ? data['summary'] : {}
-        ) as { prUrl?: string };
-        const task = (
-          typeof data['task'] === 'object' && data['task'] ? data['task'] : {}
-        ) as {
-          title?: string;
-          key?: string;
-          provider?: string;
-        };
-        const taskRef = (
-          typeof data['taskRef'] === 'object' && data['taskRef'] ? data['taskRef'] : {}
-        ) as { id?: string; raw?: string; provider?: string };
-        const convergence = deriveConvergenceFromSnapshot(snapshot);
-        const status = deriveRunListStatus(run.status, convergence.status);
-        const taskTitle = task.title ?? taskRef.raw;
-        const taskKey = task.key ?? taskRef.id;
-        const taskProvider = task.provider ?? taskRef.provider;
-        const updatedAt = run.updatedAt;
-        const startedAt = run.startedAt;
-        const prUrl = summary.prUrl;
-
-        const runEntry: RunListEntry = {
-          runId,
-          status,
-          phase: run.phase ?? 'unknown',
-          ...(typeof taskTitle === 'string' ? { taskTitle } : {}),
-          ...(typeof taskKey === 'string' ? { taskKey } : {}),
-          ...(typeof taskProvider === 'string' ? { taskProvider } : {}),
-          ...(typeof updatedAt === 'string' ? { updatedAt } : {}),
-          ...(typeof startedAt === 'string' ? { startedAt } : {}),
-          ...(typeof prUrl === 'string' ? { prUrl } : {}),
-          convergence: { status: convergence.status, reason: convergence.message },
-        };
-
-        runs.push(runEntry);
-      }
-
-      runs.sort((a, b) => (b.updatedAt ?? '').localeCompare(a.updatedAt ?? ''));
-
-      const statusFilter = parseListFlag(options.status);
-      const phaseFilter = parseListFlag(options.phase);
-      const sourceFilter = parseListFlag(options.source);
-
-      let filtered = runs;
-      if (statusFilter) {
-        filtered = filtered.filter((run) => statusFilter.includes(run.status));
-      }
-      if (phaseFilter) {
-        filtered = filtered.filter((run) =>
-          phaseFilter.includes((run.phase ?? 'unknown').toLowerCase()),
-        );
-      }
-      if (sourceFilter) {
-        filtered = filtered.filter((run) =>
-          sourceFilter.includes((run.taskProvider ?? 'unknown').toLowerCase()),
-        );
-      }
-
-      const total = runs.length;
-      const filteredTotal = filtered.length;
-      const limit = Math.max(1, parseNumberFlag(options.limit) ?? 20);
-      const offset = Math.max(0, parseNumberFlag(options.offset) ?? 0);
-      const paged = filtered.slice(offset, offset + limit);
-
-      const format = (options.format ?? 'table').toLowerCase();
-      const useJson = Boolean(options.json) || format === 'json';
-      if (!['table', 'minimal', 'json'].includes(format)) {
-        throw new SilvanError({
-          code: 'run.list.invalid_format',
-          message: `Unknown format: ${format}`,
-          userMessage: `Unknown format: ${format}`,
-          kind: 'validation',
-          nextSteps: ['Use --format table, minimal, or json.'],
-        });
-      }
-
-      if (useJson) {
-        const jsonRuns = paged.map((run) => ({
-          id: run.runId,
-          status: run.status,
-          phase: run.phase,
-          task: {
-            title: run.taskTitle ?? 'Untitled',
-            source: run.taskProvider ?? null,
-            key: run.taskKey ?? null,
-          },
-          updatedAt: run.updatedAt ?? null,
-          startedAt: run.startedAt ?? null,
-          ...(run.prUrl ? { pr: { url: run.prUrl } } : {}),
-          ...(run.convergence
-            ? {
-                convergence: {
-                  status: run.convergence.status,
-                  reason: run.convergence.reason,
-                },
-              }
-            : {}),
-        }));
-        const firstRun = paged[0];
-        const nextSteps = firstRun
-          ? [
-              `silvan run inspect ${firstRun.runId}`,
-              `silvan run status ${firstRun.runId}`,
-            ]
-          : ['silvan task start "Your task"'];
-        await emitJsonSuccess({
-          command: 'run list',
-          data: {
-            total,
-            filtered: filteredTotal,
-            showing: paged.length,
-            runs: jsonRuns,
-          },
-          nextSteps,
-          repoRoot: repo.projectRoot,
-        });
-        return;
-      }
-
-      if (options.quiet) {
-        return;
-      }
-
-      if (format === 'minimal') {
-        console.log(renderRunListMinimal(paged));
-        return;
-      }
-
-      const output = renderRunListTable(paged, {
-        total,
-        filteredTotal,
-        showing: paged.length,
-        limit,
-        offset,
-        showSource: Boolean(options.showSource),
-        filters: {
-          ...(statusFilter ? { status: statusFilter } : {}),
-          ...(phaseFilter ? { phase: phaseFilter } : {}),
-          ...(sourceFilter ? { source: sourceFilter } : {}),
-        },
-      });
-      console.log(output);
-    },
-  );
-
-cli
-  .command('analytics', 'Run analytics and success reporting')
-  .option('--since <time>', 'Filter runs starting after this time (e.g., 7d)')
-  .option('--until <time>', 'Filter runs starting before this time (e.g., 2025-01-31)')
-  .option('--provider <provider>', 'Filter by task provider (comma-separated)')
-  .option('--repo <repo>', 'Filter by repository (comma-separated)')
-  .action(
-    async (
-      options: CliOptions & {
-        since?: string;
-        until?: string;
-        provider?: string;
-        repo?: string;
-      },
-    ) => {
-      const configResult = await loadConfig(buildConfigOverrides(options), {
-        cwd: process.cwd(),
-      });
-      const repo = await detectRepoContext({ cwd: configResult.projectRoot });
-      const state = await initStateStore(repo.projectRoot, {
-        lock: false,
-        mode: configResult.config.state.mode,
-        ...(configResult.config.state.root
-          ? { root: configResult.config.state.root }
-          : {}),
-        metadataRepoRoot: repo.gitRoot,
-      });
-
-      const sinceInput = options.since?.trim();
-      const untilInput = options.until?.trim();
-      const since = sinceInput ? parseTimeInput(sinceInput) : undefined;
-      if (sinceInput && !since) {
-        throw new SilvanError({
-          code: 'analytics.invalid_since',
-          message: `Invalid --since value: ${sinceInput}`,
-          userMessage: `Invalid --since value: ${sinceInput}`,
-          kind: 'validation',
-          nextSteps: ['Use a duration like 7d or an ISO timestamp.'],
-        });
-      }
-      const until = untilInput ? parseTimeInput(untilInput) : undefined;
-      if (untilInput && !until) {
-        throw new SilvanError({
-          code: 'analytics.invalid_until',
-          message: `Invalid --until value: ${untilInput}`,
-          userMessage: `Invalid --until value: ${untilInput}`,
-          kind: 'validation',
-          nextSteps: ['Use a duration like 7d or an ISO timestamp.'],
-        });
-      }
-      if (since && until && since > until) {
-        throw new SilvanError({
-          code: 'analytics.invalid_range',
-          message: `Invalid time range: ${since} is after ${until}`,
-          userMessage: 'Invalid time range: --since must be before --until.',
-          kind: 'validation',
-          nextSteps: ['Swap the range or remove --until.'],
-        });
-      }
-
-      const providerFilter = parseListFlag(options.provider);
-      const repoFilter = parseListFlag(options.repo);
-      const report = await buildAnalyticsReport({
-        state,
-        filters: {
-          ...(since ? { since } : {}),
-          ...(until ? { until } : {}),
-          ...(providerFilter ? { providers: providerFilter } : {}),
-          ...(repoFilter ? { repos: repoFilter } : {}),
-        },
-      });
-      const nextSteps = buildAnalyticsNextSteps(report);
-
-      if (options.json) {
-        await emitJsonSuccess({
-          command: 'analytics',
-          data: report,
-          nextSteps,
-          repoRoot: repo.projectRoot,
-        });
-        return;
-      }
-
-      if (options.quiet) {
-        return;
-      }
-
-      console.log(renderAnalyticsReport(report));
-    },
-  );
-
-cli
-  .command('models recommend', 'Recommend cognition models based on recent runs')
-  .option('--min-samples <n>', 'Minimum samples per model before recommending')
-  .option('--lookback-days <n>', 'Lookback window in days')
-  .action(
-    async (options: CliOptions & { minSamples?: string; lookbackDays?: string }) => {
-      const configResult = await loadConfig(buildConfigOverrides(options), {
-        cwd: process.cwd(),
-      });
-      const repo = await detectRepoContext({ cwd: configResult.projectRoot });
-      const state = await initStateStore(repo.projectRoot, {
-        lock: false,
-        mode: configResult.config.state.mode,
-        ...(configResult.config.state.root
-          ? { root: configResult.config.state.root }
-          : {}),
-        metadataRepoRoot: repo.gitRoot,
-      });
-
-      const minSamples = parseNumberFlag(options.minSamples);
-      const lookbackDays = parseNumberFlag(options.lookbackDays);
-      const config: Config = {
-        ...configResult.config,
-        ai: {
-          ...configResult.config.ai,
-          cognition: {
-            ...configResult.config.ai.cognition,
-            routing: {
-              ...configResult.config.ai.cognition.routing,
-              ...(minSamples ? { minSamples } : {}),
-              ...(lookbackDays ? { lookbackDays } : {}),
-            },
-          },
-        },
-      };
-
-      const report = await buildModelRoutingReport({ state, config });
-      const nextSteps = ['silvan models benchmark --models <model-a,model-b>'];
-
-      if (options.json) {
-        await emitJsonSuccess({
-          command: 'models recommend',
-          data: report,
-          nextSteps,
-          repoRoot: repo.projectRoot,
-        });
-        return;
-      }
-
-      if (options.quiet) {
-        return;
-      }
-
-      console.log(
-        renderModelRoutingReport({
-          report,
-          autoApply: config.ai.cognition.routing.autoApply,
-          minSamples: config.ai.cognition.routing.minSamples,
-        }),
-      );
-    },
-  );
-
-cli
-  .command('models benchmark', 'Benchmark cognition models with a sample plan')
-  .option('--models <models>', 'Comma-separated list of models to compare')
-  .action(async (options: CliOptions & { models?: string }) => {
-    const configResult = await loadConfig(buildConfigOverrides(options), {
-      cwd: process.cwd(),
-    });
-    const repo = await detectRepoContext({ cwd: configResult.projectRoot });
-    const state = await initStateStore(repo.projectRoot, {
-      lock: false,
-      mode: configResult.config.state.mode,
-      ...(configResult.config.state.root ? { root: configResult.config.state.root } : {}),
-      metadataRepoRoot: repo.gitRoot,
-    });
-
-    const explicitModels = parseModelList(options.models);
-    const models = explicitModels ?? collectBenchmarkModels(configResult.config);
-    const uniqueModels = Array.from(new Set(models));
-    if (uniqueModels.length < 2) {
-      throw new SilvanError({
-        code: 'models.benchmark.insufficient_models',
-        message: 'Provide at least two models to benchmark.',
-        userMessage: 'Provide at least two models to benchmark.',
-        kind: 'validation',
-        nextSteps: [
-          'Pass --models model-a,model-b',
-          'Or set ai.models.default and ai.cognition.modelByTask.plan in config',
-        ],
-      });
-    }
-
-    const report = await benchmarkCognitionModels({
-      state,
-      config: configResult.config,
-      repoRoot: repo.projectRoot,
-      models: uniqueModels,
-    });
-    const nextSteps = ['silvan models recommend'];
-
-    if (options.json) {
-      await emitJsonSuccess({
-        command: 'models benchmark',
-        data: report,
-        nextSteps,
-        repoRoot: repo.projectRoot,
-      });
-      return;
-    }
-
-    if (options.quiet) {
-      return;
-    }
-
-    console.log(renderModelBenchmarkReport(report));
-  });
-
-cli
-  .command('logs <runId>', 'Show audit log for a run')
-  .option('--tail <n>', 'Show the last N events')
-  .action(async (runId: string, options: CliOptions & { tail?: string }) => {
-    const configResult = await loadConfig(buildConfigOverrides(options), {
-      cwd: process.cwd(),
-    });
-    const repo = await detectRepoContext({ cwd: configResult.projectRoot });
-    const state = await initStateStore(repo.projectRoot, {
-      lock: false,
-      mode: configResult.config.state.mode,
-      ...(configResult.config.state.root ? { root: configResult.config.state.root } : {}),
-      metadataRepoRoot: repo.gitRoot,
-    });
-    const logPath = join(state.auditDir, `${runId}.jsonl`);
-    const logFile = Bun.file(logPath);
-    if (!(await logFile.exists())) {
-      throw new SilvanError({
-        code: 'audit_log.not_found',
-        message: `Audit log not found for run ${runId} (${logPath}).`,
-        userMessage: `Audit log not found for run ${runId}.`,
-        kind: 'not_found',
-        details: { runId, logPath },
-        nextSteps: ['Check the run ID with `silvan run list`.'],
-      });
-    }
-
-    const content = await logFile.text();
-    const lines = content
-      .split('\n')
-      .map((line) => line.trim())
-      .filter((line) => line.length > 0);
-    const limit = parseNumberFlag(options.tail);
-    const selected = limit ? lines.slice(-limit) : lines;
-    const events = selected
-      .map((line) => parseAuditEvent(line))
-      .filter((event): event is Event => event !== null);
-
-    if (options.json) {
-      await emitJsonSuccess({
-        command: 'logs',
-        data: { runId, events },
-        repoRoot: repo.projectRoot,
-      });
-      return;
-    }
-
-    for (const event of events) {
-      console.log(formatAuditEvent(event));
-    }
-  });
-
-cli
-  .command('run inspect <runId>', 'Inspect a run snapshot')
-  .action(async (runId: string, options: CliOptions) => {
-    const configResult = await loadConfig(buildConfigOverrides(options), {
-      cwd: process.cwd(),
-    });
-    const repo = await detectRepoContext({ cwd: configResult.projectRoot });
-    const state = await initStateStore(repo.projectRoot, {
-      lock: false,
-      mode: configResult.config.state.mode,
-      ...(configResult.config.state.root ? { root: configResult.config.state.root } : {}),
-      metadataRepoRoot: repo.gitRoot,
-    });
-    const snapshot = await state.readRunState(runId);
-    if (!snapshot) {
-      throw new Error(`Run not found: ${runId}`);
-    }
-    if (options.json) {
-      await emitJsonSuccess({
-        command: 'run inspect',
-        data: { snapshot },
-        repoRoot: repo.projectRoot,
-        runId,
-      });
-      return;
-    }
-    if (options.quiet) {
-      return;
-    }
-    const data = snapshot.data as Record<string, unknown>;
-    const task = (
-      typeof data['task'] === 'object' && data['task'] ? data['task'] : {}
-    ) as {
-      id?: string;
-      key?: string;
-      title?: string;
-      provider?: string;
-      acceptanceCriteria?: string[];
-    };
-    const run = (typeof data['run'] === 'object' && data['run'] ? data['run'] : {}) as {
-      status?: string;
-      phase?: string;
-      updatedAt?: string;
-    };
-    const acCount = Array.isArray(task.acceptanceCriteria)
-      ? task.acceptanceCriteria.length
-      : 0;
-
-    const lines: string[] = [];
-    lines.push(renderSectionHeader('Run snapshot', { width: 60, kind: 'minor' }));
-    lines.push(
-      ...formatKeyValues(
-        [
-          ['Run ID', snapshot.runId],
-          ['Status', formatStatusLabel(run.status ?? 'unknown')],
-          ['Phase', run.phase ?? 'unknown'],
-          ['Updated', run.updatedAt ?? 'n/a'],
-        ],
-        { labelWidth: 12 },
-      ),
-    );
-    lines.push(
-      ...formatKeyValues([['State file', `${snapshot.runId}.json`]], { labelWidth: 12 }),
-    );
-
-    if (task.id || task.title) {
-      lines.push('');
-      lines.push(renderSectionHeader('Task', { width: 60, kind: 'minor' }));
-      lines.push(
-        ...formatKeyValues(
-          [
-            ['Ref', task.key ?? task.id ?? 'unknown'],
-            ['Title', task.title ?? 'Untitled'],
-            ['Provider', task.provider ?? 'unknown'],
-            ['Criteria', `${acCount} item(s)`],
-          ],
-          { labelWidth: 12 },
-        ),
-      );
-    }
-
-    lines.push(
-      renderNextSteps([`silvan run status ${runId}`, `silvan run explain ${runId}`]),
-    );
-    console.log(lines.join('\n'));
-  });
-
-cli
-  .command('run status <runId>', 'Show convergence status for a run')
-  .action(async (runId: string, options: CliOptions) => {
-    const { snapshot, repoRoot } = await loadRunSnapshotForCli(runId, options);
-    const convergence = deriveConvergenceFromSnapshot(snapshot);
-    if (options.json) {
-      await emitJsonSuccess({
-        command: 'run status',
-        data: {
-          runId,
-          status: convergence.status,
-          reasonCode: convergence.reasonCode,
-          message: convergence.message,
-          nextActions: convergence.nextActions,
-          blockingArtifacts: convergence.blockingArtifacts ?? [],
-        },
-        repoRoot,
-        runId,
-      });
-      return;
-    }
-    if (options.quiet) {
-      return;
-    }
-    const lines: string[] = [];
-    lines.push(renderSectionHeader('Run status', { width: 60, kind: 'minor' }));
-    lines.push(
-      ...formatKeyValues(
-        [
-          ['Run ID', runId],
-          ['Status', formatStatusLabel(convergence.status)],
-          ['Reason', convergence.reasonCode],
-        ],
-        { labelWidth: 12 },
-      ),
-    );
-    lines.push(`Message: ${convergence.message}`);
-    if (convergence.blockingArtifacts?.length) {
-      lines.push(
-        ...formatKeyList(
-          'Blocking',
-          `${convergence.blockingArtifacts.length} artifact(s)`,
-          convergence.blockingArtifacts,
-          { labelWidth: 12 },
-        ),
-      );
-    }
-    if (convergence.nextActions.length) {
-      lines.push(
-        ...formatKeyList(
-          'Next actions',
-          `${convergence.nextActions.length} action(s)`,
-          convergence.nextActions,
-          { labelWidth: 12 },
-        ),
-      );
-    }
-    lines.push(
-      renderNextSteps([`silvan run explain ${runId}`, `silvan run resume ${runId}`]),
-    );
-    console.log(lines.join('\n'));
-  });
-
-cli
-  .command('run explain <runId>', 'Explain why a run is waiting or blocked')
-  .action(async (runId: string, options: CliOptions) => {
-    const { snapshot, repoRoot } = await loadRunSnapshotForCli(runId, options);
-    const data = snapshot.data as Record<string, unknown>;
-    const run = (typeof data['run'] === 'object' && data['run'] ? data['run'] : {}) as {
-      status?: string;
-      phase?: string;
-    };
-    const steps = (
-      typeof data['steps'] === 'object' && data['steps'] ? data['steps'] : {}
-    ) as Record<string, { status?: string; endedAt?: string }>;
-    const summary = (
-      typeof data['summary'] === 'object' && data['summary'] ? data['summary'] : {}
-    ) as {
-      prUrl?: string;
-      ci?: string;
-      unresolvedReviewCount?: number;
-      blockedReason?: string;
-    };
-    const reviewPriority = Array.isArray(data['reviewThreadPriority'])
-      ? (data['reviewThreadPriority'] as Array<{
-          threadId?: string;
-          severity?: string;
-          summary?: string;
-          path?: string | null;
-          line?: number | null;
-          isOutdated?: boolean;
-        }>)
-      : [];
-    const reviewerSuggestions = (data['reviewerSuggestions'] as
-      | {
-          users?: string[];
-          teams?: string[];
-          error?: string;
-        }
-      | undefined) ?? { users: [], teams: [] };
-    const reviewResponseSummary =
-      (data['reviewResponseSummary'] as
-        | {
-            requestedAt?: string;
-            reviewers?: string[];
-            respondedReviewers?: string[];
-            pendingReviewers?: string[];
-            avgResponseHours?: number;
-          }
-        | undefined) ?? {};
-    const verificationAssist = data['verificationAssistSummary'] as
-      | {
-          summary?: string;
-          steps?: string[];
-          context?: string;
-          commands?: string[];
-        }
-      | undefined;
-    const localGate = (data['localGateSummary'] as
-      | { ok?: boolean; blockers?: number; warnings?: number }
-      | undefined) ?? { ok: undefined };
-    const convergence = deriveConvergenceFromSnapshot(snapshot);
-    const lastStep = findLastSuccessfulStep(steps);
-
-    if (options.json) {
-      await emitJsonSuccess({
-        command: 'run explain',
-        data: {
-          runId,
-          run: { status: run.status, phase: run.phase },
-          convergence,
-          lastSuccessfulStep: lastStep ?? null,
-          summaries: {
-            ci: summary.ci ?? null,
-            unresolvedReviewCount: summary.unresolvedReviewCount ?? null,
-            blockedReason: summary.blockedReason ?? null,
-            localGate,
-            verificationAssist: verificationAssist ?? null,
-          },
-          review: {
-            priorityThreads: reviewPriority,
-            reviewerSuggestions,
-            responseSummary: reviewResponseSummary,
-          },
-        },
-        repoRoot,
-        runId,
-      });
-      return;
-    }
-    if (options.quiet) {
-      return;
-    }
-    const lines: string[] = [];
-    lines.push(renderSectionHeader('Run explanation', { width: 60, kind: 'minor' }));
-    lines.push(
-      ...formatKeyValues(
-        [
-          ['Run ID', runId],
-          ['Status', formatStatusLabel(run.status ?? 'unknown')],
-          ['Phase', run.phase ?? 'unknown'],
-        ],
-        { labelWidth: 12 },
-      ),
-    );
-    lines.push(
-      ...formatKeyValues(
-        [
-          ['Convergence', formatStatusLabel(convergence.status)],
-          ['Reason', convergence.reasonCode],
-        ],
-        { labelWidth: 12 },
-      ),
-    );
-    lines.push(`Message: ${convergence.message}`);
-    if (lastStep) {
-      lines.push(`Last successful step: ${lastStep}`);
-    }
-    if (summary.prUrl) {
-      lines.push(`PR: ${summary.prUrl}`);
-    }
-    if (summary.ci) {
-      lines.push(`CI: ${summary.ci}`);
-    }
-    if (typeof summary.unresolvedReviewCount === 'number') {
-      lines.push(`Unresolved review comments: ${summary.unresolvedReviewCount}`);
-    }
-    if (summary.blockedReason) {
-      lines.push(`Blocked reason: ${summary.blockedReason}`);
-    }
-    if (reviewPriority.length > 0) {
-      lines.push('');
-      lines.push(renderSectionHeader('Review priorities', { width: 60, kind: 'minor' }));
-      const list = reviewPriority.map((thread) => {
-        const severity = thread.severity ?? 'unknown';
-        const summaryText = thread.summary ?? thread.threadId ?? 'Thread';
-        const location = thread.path
-          ? `${thread.path}${thread.line ? `:${thread.line}` : ''}`
-          : 'unknown location';
-        return `${severity.toUpperCase()}: ${summaryText} (${location})`;
-      });
-      lines.push(
-        ...formatKeyList('Threads', `${reviewPriority.length} item(s)`, list, {
-          labelWidth: 12,
-        }),
-      );
-    }
-    if (
-      reviewerSuggestions.error ||
-      (reviewerSuggestions.users?.length ?? 0) > 0 ||
-      (reviewerSuggestions.teams?.length ?? 0) > 0
-    ) {
-      lines.push('');
-      lines.push(
-        renderSectionHeader('Reviewer suggestions', { width: 60, kind: 'minor' }),
-      );
-      if (reviewerSuggestions.users?.length) {
-        lines.push(
-          ...formatKeyList(
-            'Users',
-            `${reviewerSuggestions.users.length} suggested`,
-            reviewerSuggestions.users,
-            { labelWidth: 12 },
-          ),
-        );
-      }
-      if (reviewerSuggestions.teams?.length) {
-        lines.push(
-          ...formatKeyList(
-            'Teams',
-            `${reviewerSuggestions.teams.length} suggested`,
-            reviewerSuggestions.teams,
-            { labelWidth: 12 },
-          ),
-        );
-      }
-      if (reviewerSuggestions.error) {
-        lines.push(`Notes: ${reviewerSuggestions.error}`);
-      }
-    }
-    if (reviewResponseSummary.reviewers?.length) {
-      const respondedCount = reviewResponseSummary.respondedReviewers?.length ?? 0;
-      const pendingCount = reviewResponseSummary.pendingReviewers?.length ?? 0;
-      lines.push('');
-      lines.push(renderSectionHeader('Reviewer responses', { width: 60, kind: 'minor' }));
-      const responseDetails: Array<[string, string]> = [
-        ['Requested', `${reviewResponseSummary.reviewers.length} reviewer(s)`],
-        ['Responded', `${respondedCount} reviewer(s)`],
-        ['Pending', `${pendingCount} reviewer(s)`],
-      ];
-      if (typeof reviewResponseSummary.avgResponseHours === 'number') {
-        responseDetails.push([
-          'Avg response',
-          `${reviewResponseSummary.avgResponseHours.toFixed(1)}h`,
-        ]);
-      }
-      lines.push(...formatKeyValues(responseDetails, { labelWidth: 12 }));
-    }
-    if (verificationAssist?.summary || verificationAssist?.steps?.length) {
-      lines.push('');
-      lines.push(
-        renderSectionHeader('Verification assist', { width: 60, kind: 'minor' }),
-      );
-      const details: Array<[string, string]> = [];
-      if (verificationAssist.context) {
-        details.push(['Context', verificationAssist.context]);
-      }
-      if (verificationAssist.commands?.length) {
-        details.push(['Commands', verificationAssist.commands.join(', ')]);
-      }
-      if (verificationAssist.summary) {
-        details.push(['Summary', verificationAssist.summary]);
-      }
-      if (details.length > 0) {
-        lines.push(...formatKeyValues(details, { labelWidth: 12 }));
-      }
-      if (verificationAssist.steps?.length) {
-        lines.push(
-          ...formatKeyList(
-            'Steps',
-            `${verificationAssist.steps.length} action(s)`,
-            verificationAssist.steps,
-            { labelWidth: 12 },
-          ),
-        );
-      }
-    }
-    if (localGate.ok === false) {
-      lines.push(
-        `Local gate: ${localGate.blockers ?? 0} blockers, ${localGate.warnings ?? 0} warnings`,
-      );
-    }
-    if (convergence.blockingArtifacts?.length) {
-      lines.push(
-        ...formatKeyList(
-          'Blocking',
-          `${convergence.blockingArtifacts.length} artifact(s)`,
-          convergence.blockingArtifacts,
-          { labelWidth: 12 },
-        ),
-      );
-    }
-    if (convergence.nextActions.length) {
-      lines.push(
-        ...formatKeyList(
-          'Next actions',
-          `${convergence.nextActions.length} action(s)`,
-          convergence.nextActions,
-          { labelWidth: 12 },
-        ),
-      );
-    }
-    lines.push(
-      renderNextSteps([`silvan run resume ${runId}`, `silvan run status ${runId}`]),
-    );
-    console.log(lines.join('\n'));
-  });
-
-cli
-  .command('learning show <runId>', 'Show learning notes for a run')
-  .action(async (runId: string, options: CliOptions) => {
-    const { snapshot, repoRoot } = await loadRunSnapshotForCli(runId, options);
-    const data = snapshot.data as Record<string, unknown>;
-    const artifactsIndex =
-      typeof data['artifactsIndex'] === 'object' && data['artifactsIndex']
-        ? (data['artifactsIndex'] as Record<string, Record<string, unknown>>)
-        : undefined;
-    const notesEntry =
-      artifactsIndex?.['learning.notes']?.['notes'] ??
-      artifactsIndex?.['learning.notes']?.['data'];
-    if (!notesEntry || typeof notesEntry !== 'object') {
-      throw new Error('Learning notes not found for this run.');
-    }
-    if (!isArtifactEntry(notesEntry)) {
-      throw new Error('Learning notes artifact entry is invalid.');
-    }
-    const content = await readArtifact({ entry: notesEntry });
-    if (options.json) {
-      await emitJsonSuccess({
-        command: 'learning show',
-        data: { runId, kind: notesEntry.kind, content },
-        repoRoot,
-        runId,
-      });
-      return;
-    }
-    if (notesEntry.kind === 'text') {
-      console.log(content);
-      return;
-    }
-    console.log(JSON.stringify(content, null, 2));
-  });
-
-function isArtifactEntry(value: unknown): value is ArtifactEntry {
-  if (!value || typeof value !== 'object') return false;
-  const record = value as Record<string, unknown>;
-  return (
-    typeof record['path'] === 'string' &&
-    typeof record['stepId'] === 'string' &&
-    typeof record['name'] === 'string' &&
-    typeof record['digest'] === 'string' &&
-    typeof record['updatedAt'] === 'string' &&
-    (record['kind'] === 'json' || record['kind'] === 'text')
-  );
-}
-
-cli
-  .command('learning review', 'Review pending learning notes')
-  .option('--approve <runIds>', 'Approve pending learnings (comma-separated run IDs)')
-  .option('--reject <runIds>', 'Reject pending learnings (comma-separated run IDs)')
-  .option('--all', 'Apply action to all pending learnings')
-  .action((options: CliOptions & { approve?: string; reject?: string; all?: boolean }) =>
-    withCliContext(options, options.json ? 'json' : 'headless', async (ctx) => {
-      const logger = createCliLogger(ctx);
-      const requests = await listLearningRequests({ state: ctx.state });
-      const pending = requests.filter((request) => request.status === 'pending');
-
-      const approveIds = parseCsvFlag(options.approve);
-      const rejectIds = parseCsvFlag(options.reject);
-      const action = approveIds ? 'approve' : rejectIds ? 'reject' : undefined;
-
-      if (approveIds && rejectIds) {
-        throw new SilvanError({
-          code: 'learning.review.conflict',
-          message: 'Choose either --approve or --reject.',
-          userMessage: 'Choose either --approve or --reject.',
-          kind: 'validation',
-          nextSteps: ['Run `silvan learning review --approve <runId>` or `--reject`.'],
-        });
-      }
-
-      if (options.all && !action) {
-        throw new SilvanError({
-          code: 'learning.review.missing_action',
-          message: 'Specify --approve or --reject when using --all.',
-          userMessage: 'Specify --approve or --reject when using --all.',
-          kind: 'validation',
-          nextSteps: ['Run `silvan learning review --approve --all`.'],
-        });
-      }
-
-      if (!action) {
-        if (options.json) {
-          await emitJsonSuccess({
-            command: 'learning review',
-            data: {
-              pending: pending.map((request) => ({
-                runId: request.runId,
-                summary: request.summary,
-                confidence: request.confidence,
-                threshold: request.threshold,
-                reason: request.reason ?? null,
-                createdAt: request.createdAt,
-              })),
-            },
-            repoRoot: ctx.repo.repoRoot,
-            runId: ctx.runId,
-          });
-          return;
-        }
-        if (options.quiet) {
-          return;
-        }
-        if (pending.length === 0) {
-          await logger.info('No pending learning notes.');
-          return;
-        }
-        const lines: string[] = [];
-        lines.push(renderSectionHeader('Pending learning notes', { width: 60 }));
-        lines.push(
-          ...formatKeyValues([['Pending', `${pending.length} item(s)`]], {
-            labelWidth: 12,
-          }),
-        );
-        const summaries = pending.map((request) => {
-          const confidence = `${Math.round(request.confidence * 100)}%`;
-          return `${request.runId} (${confidence}): ${request.summary}`;
-        });
-        lines.push(
-          ...formatKeyList('Items', `${pending.length} item(s)`, summaries, {
-            labelWidth: 12,
-          }),
-        );
-        lines.push(
-          renderNextSteps([
-            'silvan learning review --approve <runId>',
-            'silvan learning review --reject <runId>',
-          ]),
-        );
-        await logger.info(lines.join('\n'));
-        return;
-      }
-
-      const targetIds = options.all
-        ? pending.map((request) => request.id)
-        : action === 'approve'
-          ? (approveIds ?? [])
-          : (rejectIds ?? []);
-
-      const selection = pending.filter((request) => targetIds.includes(request.id));
-      const missing = targetIds.filter(
-        (id) => !pending.some((request) => request.id === id),
-      );
-
-      const results: Array<{ runId: string; status: string; message?: string }> = [];
-      for (const request of selection) {
-        try {
-          if (action === 'approve') {
-            const result = await applyLearningRequestForCli(ctx, request);
-            results.push({
-              runId: request.runId,
-              status: 'applied',
-              ...(result.commitSha ? { message: `commit ${result.commitSha}` } : {}),
-            });
-          } else {
-            await rejectLearningRequestForCli(ctx, request);
-            results.push({ runId: request.runId, status: 'rejected' });
-          }
-        } catch (error) {
-          const message =
-            error instanceof Error ? error.message : 'Learning review failed';
-          results.push({ runId: request.runId, status: 'failed', message });
-        }
-      }
-
-      const failed = results.filter((result) => result.status === 'failed');
-      if (options.json) {
-        await emitJsonResult({
-          command: 'learning review',
-          success: failed.length === 0,
-          data: {
-            action,
-            processed: results.length,
-            failed: failed.length,
-            missing,
-            results,
-          },
-          ...(failed.length > 0
-            ? {
-                error: {
-                  code: 'learning.review.failed',
-                  message: 'One or more learning review actions failed.',
-                  details: { failed },
-                  suggestions: ['Inspect failures, then re-run the command.'],
-                },
-              }
-            : {}),
-          repoRoot: ctx.repo.repoRoot,
-          runId: ctx.runId,
-        });
-        return;
-      }
-
-      if (options.quiet) {
-        return;
-      }
-      const lines: string[] = [];
-      const title =
-        failed.length > 0
-          ? 'Learning review completed with errors'
-          : 'Learning review completed';
-      lines.push(renderSectionHeader(title, { width: 60, kind: 'minor' }));
-      lines.push(
-        ...formatKeyValues(
-          [
-            ['Action', action],
-            ['Processed', `${results.length} item(s)`],
-            ['Failed', `${failed.length} item(s)`],
-          ],
-          { labelWidth: 12 },
-        ),
-      );
-      if (missing.length > 0) {
-        lines.push(
-          ...formatKeyList('Missing', `${missing.length} item(s)`, missing, {
-            labelWidth: 12,
-          }),
-        );
-      }
-      if (failed.length > 0) {
-        lines.push(
-          ...formatKeyList(
-            'Failures',
-            `${failed.length} item(s)`,
-            failed.map((result) => `${result.runId}: ${result.message}`),
-            { labelWidth: 12 },
-          ),
-        );
-      }
-      lines.push(renderNextSteps(['silvan learning review', 'silvan run list']));
-      await logger.info(lines.join('\n'));
-      if (failed.length > 0) {
-        process.exitCode = 1;
-      }
-    }),
-  );
-
-cli
-  .command('learning rollback <runId>', 'Rollback applied learning updates')
-  .action((runId: string, options: CliOptions) =>
-    withCliContext(options, options.json ? 'json' : 'headless', async (ctx) => {
-      const logger = createCliLogger(ctx);
-      const result = await rollbackLearningRequestForCli(ctx, runId);
-
-      if (options.json) {
-        await emitJsonSuccess({
-          command: 'learning rollback',
-          data: result,
-          repoRoot: ctx.repo.repoRoot,
-          runId: ctx.runId,
-        });
-        return;
-      }
-      if (options.quiet) {
-        return;
-      }
-      await logger.info(
-        renderSuccessSummary({
-          title: 'Learning rollback complete',
-          details: [
-            ['Run ID', runId],
-            ['Revert', result.revertSha ?? 'unknown'],
-          ],
-          nextSteps: [`silvan run status ${runId}`, 'silvan learning review'],
-        }),
-      );
-    }),
-  );
-
-async function applyLearningRequestForCli(
-  ctx: RunContext,
-  request: LearningRequest,
-): Promise<{ appliedTo: string[]; commitSha?: string }> {
-  const snapshot = await ctx.state.readRunState(request.runId);
-  if (!snapshot) {
-    throw new SilvanError({
-      code: 'learning.review.run_missing',
-      message: `Run not found: ${request.runId}`,
-      userMessage: `Run not found: ${request.runId}`,
-      kind: 'not_found',
-      nextSteps: ['Run `silvan run list` to check available runs.'],
-    });
-  }
-  const data = snapshot.data as Record<string, unknown>;
-  const worktreeRoot = await resolveLearningWorktreeRoot(ctx, data, request.runId);
-  const targetCheck = evaluateLearningTargets({
-    targets: request.targets,
-    worktreeRoot,
-  });
-  if (!targetCheck.ok) {
-    throw new SilvanError({
-      code: 'learning.review.unsafe_targets',
-      message: targetCheck.reasons.join('; '),
-      userMessage: 'Learning targets are not safe to apply automatically.',
-      kind: 'validation',
-      nextSteps: [
-        'Update learning.targets to point at documentation files.',
-        'Re-run `silvan learning review` after fixing the targets.',
-      ],
-    });
-  }
-  const applyResult = await applyLearningNotes({
-    runId: request.runId,
-    worktreeRoot,
-    notes: request.notes,
-    targets: request.targets,
-  });
-  const commitResult = await commitLearningRequestForCli({
-    ctx,
-    runId: request.runId,
-    worktreeRoot,
-    appliedTo: applyResult.appliedTo,
-  });
-  const appliedAt = new Date().toISOString();
-
-  await ctx.state.updateRunState(request.runId, (prev) => ({
-    ...prev,
-    learningSummary: {
-      ...(typeof prev['learningSummary'] === 'object' && prev['learningSummary']
-        ? (prev['learningSummary'] as Record<string, unknown>)
-        : {}),
-      status: 'applied',
-      appliedAt,
-      appliedTo: applyResult.appliedTo,
-      commitSha: commitResult.sha,
-      autoApplied: false,
-    },
-  }));
-
-  await writeLearningRequest({
-    state: ctx.state,
-    request: {
-      ...request,
-      status: 'applied',
-      updatedAt: appliedAt,
-      appliedAt,
-      appliedTo: applyResult.appliedTo,
-      ...(commitResult.sha ? { commitSha: commitResult.sha } : {}),
-    },
-  });
-
-  return {
-    appliedTo: applyResult.appliedTo,
-    ...(commitResult.sha ? { commitSha: commitResult.sha } : {}),
-  };
-}
-
-async function rejectLearningRequestForCli(
-  ctx: RunContext,
-  request: LearningRequest,
-): Promise<void> {
-  const rejectedAt = new Date().toISOString();
-  await ctx.state.updateRunState(request.runId, (prev) => ({
-    ...prev,
-    learningSummary: {
-      ...(typeof prev['learningSummary'] === 'object' && prev['learningSummary']
-        ? (prev['learningSummary'] as Record<string, unknown>)
-        : {}),
-      status: 'rejected',
-      rejectedAt,
-    },
-  }));
-  await writeLearningRequest({
-    state: ctx.state,
-    request: {
-      ...request,
-      status: 'rejected',
-      updatedAt: rejectedAt,
-      rejectedAt,
-    },
-  });
-}
-
-async function rollbackLearningRequestForCli(
-  ctx: RunContext,
-  runId: string,
-): Promise<{ runId: string; commitSha?: string; revertSha?: string }> {
-  const snapshot = await ctx.state.readRunState(runId);
-  if (!snapshot) {
-    throw new SilvanError({
-      code: 'learning.rollback.missing_run',
-      message: `Run not found: ${runId}`,
-      userMessage: `Run not found: ${runId}`,
-      kind: 'not_found',
-      nextSteps: ['Run `silvan run list` to check available runs.'],
-    });
-  }
-  const data = snapshot.data as Record<string, unknown>;
-  const learningSummary = (data['learningSummary'] as Record<string, unknown>) ?? {};
-  const request = await readLearningRequest({ state: ctx.state, requestId: runId });
-  const commitSha =
-    (typeof learningSummary['commitSha'] === 'string'
-      ? learningSummary['commitSha']
-      : undefined) ?? request?.commitSha;
-
-  if (!commitSha) {
-    throw new SilvanError({
-      code: 'learning.rollback.missing_commit',
-      message: 'No learning commit found for this run.',
-      userMessage: 'No learning commit found for this run.',
-      kind: 'validation',
-      nextSteps: ['Run `silvan learning review` to see pending notes.'],
-    });
-  }
-
-  const worktreeRoot = await resolveLearningWorktreeRoot(ctx, data, runId);
-  const revert = await runGit(['revert', '--no-edit', commitSha], {
-    cwd: worktreeRoot,
-    context: { runId, repoRoot: ctx.repo.repoRoot },
-  });
-  if (revert.exitCode !== 0) {
-    throw new SilvanError({
-      code: 'learning.rollback.failed',
-      message: revert.stderr || 'Failed to revert learning commit.',
-      userMessage: 'Failed to revert the learning commit.',
-      kind: 'internal',
-      nextSteps: [
-        'Resolve conflicts in the worktree.',
-        'Run `git revert --continue` or `git revert --abort`.',
-      ],
-    });
-  }
-
-  const shaResult = await runGit(['rev-parse', 'HEAD'], {
-    cwd: worktreeRoot,
-    context: { runId, repoRoot: ctx.repo.repoRoot },
-  });
-  const revertSha = shaResult.exitCode === 0 ? shaResult.stdout.trim() : undefined;
-  const rolledBackAt = new Date().toISOString();
-
-  await ctx.state.updateRunState(runId, (prev) => ({
-    ...prev,
-    learningSummary: {
-      ...(typeof prev['learningSummary'] === 'object' && prev['learningSummary']
-        ? (prev['learningSummary'] as Record<string, unknown>)
-        : {}),
-      status: 'rolled_back',
-      rolledBackAt,
-      rollbackSha: revertSha,
-    },
-  }));
-
-  if (request) {
-    await writeLearningRequest({
-      state: ctx.state,
-      request: {
-        ...request,
-        status: 'rolled_back',
-        updatedAt: rolledBackAt,
-        rolledBackAt,
-        commitSha,
-      },
-    });
-  }
-
-  return {
-    runId,
-    ...(commitSha ? { commitSha } : {}),
-    ...(revertSha ? { revertSha } : {}),
-  };
-}
-
-async function resolveLearningWorktreeRoot(
-  ctx: RunContext,
-  data: Record<string, unknown>,
-  runId: string,
-): Promise<string> {
-  const worktree = (typeof data['worktree'] === 'object' && data['worktree']
-    ? (data['worktree'] as { path?: string })
-    : undefined) ?? { path: undefined };
-  const candidate = worktree.path ?? ctx.repo.repoRoot;
-
-  try {
-    await access(candidate);
-  } catch {
-    throw new SilvanError({
-      code: 'learning.review.worktree_missing',
-      message: `Worktree not found for run ${runId}.`,
-      userMessage: `Worktree not found for run ${runId}.`,
-      kind: 'not_found',
-      nextSteps: ['Run `silvan tree list` to locate the worktree.'],
-    });
-  }
-  return candidate;
-}
-
-async function commitLearningRequestForCli(options: {
-  ctx: RunContext;
-  runId: string;
-  worktreeRoot: string;
-  appliedTo: string[];
-}): Promise<{ committed: boolean; sha?: string }> {
-  const relativePaths = options.appliedTo
-    .map((filePath) => relative(options.worktreeRoot, filePath))
-    .filter((path) => path && !path.startsWith('..'));
-  if (relativePaths.length === 0) {
-    return { committed: false };
-  }
-  await runGit(['add', '--', ...relativePaths], {
-    cwd: options.worktreeRoot,
-    context: { runId: options.runId, repoRoot: options.ctx.repo.repoRoot },
-  });
-  const diff = await runGit(['diff', '--cached', '--quiet'], {
-    cwd: options.worktreeRoot,
-    context: { runId: options.runId, repoRoot: options.ctx.repo.repoRoot },
-  });
-  if (diff.exitCode === 0) {
-    return { committed: false };
-  }
-  const commit = await runGit(
-    ['commit', '-m', `silvan: apply learnings (${options.runId})`],
-    {
-      cwd: options.worktreeRoot,
-      context: { runId: options.runId, repoRoot: options.ctx.repo.repoRoot },
-    },
-  );
-  if (commit.exitCode !== 0) {
-    throw new Error(commit.stderr || 'Failed to commit learning updates');
-  }
-  const shaResult = await runGit(['rev-parse', 'HEAD'], {
-    cwd: options.worktreeRoot,
-    context: { runId: options.runId, repoRoot: options.ctx.repo.repoRoot },
-  });
-  const sha = shaResult.exitCode === 0 ? shaResult.stdout.trim() : undefined;
-  return { committed: true, ...(sha ? { sha } : {}) };
-}
-
-cli
-  .command('run resume <runId>', 'Resume a run using convergence rules')
-  .option('--dry-run', 'Allow only read-only tools')
-  .option('--apply', 'Allow mutating tools')
-  .option('--dangerous', 'Allow dangerous tools (requires --apply)')
-  .action(async (runId: string, options: CliOptions) => {
-    const { snapshot, repoRoot } = await loadRunSnapshotForCli(runId, options);
-    const convergence = deriveConvergenceFromSnapshot(snapshot);
-    if (convergence.status === 'converged' || convergence.status === 'aborted') {
-      if (options.json) {
-        await emitJsonSuccess({
-          command: 'run resume',
-          data: {
-            runId,
-            status: convergence.status,
-            reason: convergence.reasonCode,
-            message: 'Resume is not applicable.',
-          },
-          repoRoot,
-          runId,
-        });
-        return;
-      }
-      if (!options.quiet) {
-        console.log(
-          renderSuccessSummary({
-            title: 'Resume not applicable',
-            details: [
-              ['Run ID', runId],
-              ['Status', convergence.status],
-              ['Reason', convergence.reasonCode],
-            ],
-            nextSteps: [`silvan run status ${runId}`, 'silvan run list'],
-          }),
-        );
-      }
-      return;
-    }
-    await withCliContext(
-      options,
-      options.json ? 'json' : 'headless',
-      async (ctx) => {
-        const logger = createCliLogger(ctx);
-        await withAgentSessions(
-          Boolean(ctx.config.ai.sessions.persist),
-          async (sessions) => {
-            const runOptions = {
-              ...(options.dryRun ? { dryRun: true } : {}),
-              ...(options.apply ? { apply: true } : {}),
-              ...(options.dangerous ? { dangerous: true } : {}),
-              sessions,
-            };
-            await resumeRun(ctx, runOptions);
-          },
-        );
-        await logger.info(
-          renderSuccessSummary({
-            title: 'Run resumed',
-            details: [['Run ID', runId]],
-            nextSteps: [`silvan run status ${runId}`, `silvan run explain ${runId}`],
-          }),
-        );
-      },
-      { runId },
-    );
-  });
-
-cli
-  .command('run override <runId> <reason...>', 'Override a run gate with a reason')
-  .action(async (runId: string, reason: string[], options: CliOptions) => {
-    const message = reason.join(' ').trim();
-    if (!message) {
-      throw new Error('Override reason is required.');
-    }
-    const { state, repoRoot } = await loadRunSnapshotForCli(runId, options);
-    const entry = await writeOverrideArtifact({ state, runId, reason: message });
-    if (options.json) {
-      await emitJsonSuccess({
-        command: 'run override',
-        data: { runId, override: entry },
-        repoRoot,
-        runId,
-      });
-      return;
-    }
-    if (options.quiet) {
-      return;
-    }
-    console.log(
-      renderSuccessSummary({
-        title: 'Override recorded',
-        details: [
-          ['Run ID', runId],
-          ['Path', entry.path],
-        ],
-        nextSteps: [`silvan run resume ${runId}`, `silvan run status ${runId}`],
-      }),
-    );
-  });
-
-cli
-  .command('run abort <runId> [reason]', 'Abort a run and mark it as canceled')
-  .action(async (runId: string, reason: string | undefined, options: CliOptions) => {
-    const { state, repoRoot } = await loadRunSnapshotForCli(runId, options);
-    const entry = await markRunAborted({ state, runId, ...(reason ? { reason } : {}) });
-    if (options.json) {
-      await emitJsonSuccess({
-        command: 'run abort',
-        data: { runId, aborted: entry },
-        repoRoot,
-        runId,
-      });
-      return;
-    }
-    if (options.quiet) {
-      return;
-    }
-    console.log(
-      renderSuccessSummary({
-        title: 'Run aborted',
-        details: [
-          ['Run ID', runId],
-          ['Path', entry.path],
-        ],
-        nextSteps: ['silvan run list', `silvan run inspect ${runId}`],
-      }),
-    );
-  });
+registerRunCommands(cli, {
+  withCliContext,
+  withAgentSessions,
+  buildConfigOverrides,
+  parseListFlag,
+  parseNumberFlag,
+});
+
+registerAnalyticsCommands(cli, {
+  buildConfigOverrides,
+  parseListFlag,
+});
+
+registerModelCommands(cli, {
+  buildConfigOverrides,
+  parseNumberFlag,
+  parseModelList,
+  collectBenchmarkModels,
+});
+
+registerLogCommands(cli, {
+  buildConfigOverrides,
+  parseNumberFlag,
+});
+
+registerLearningCommands(cli, {
+  withCliContext,
+  parseCsvFlag,
+});
 
 registerUiCommands(cli, {
   withCliContext,
@@ -3416,228 +1335,14 @@ async function handleRootCommand(options: CliOptions): Promise<void> {
   console.log(renderReturningSummary(summary));
 }
 
-async function loadRunSnapshotForCli(
-  runId: string,
-  options: CliOptions,
-): Promise<{
-  repoRoot: string;
-  state: Awaited<ReturnType<typeof initStateStore>>;
-  snapshot: Awaited<ReturnType<typeof loadRunSnapshot>>;
-  config: Awaited<ReturnType<typeof loadConfig>>['config'];
-}> {
-  const configResult = await loadConfig(buildConfigOverrides(options), {
-    cwd: process.cwd(),
-  });
-  const repo = await detectRepoContext({ cwd: configResult.projectRoot });
-  const state = await initStateStore(repo.projectRoot, {
-    lock: false,
-    mode: configResult.config.state.mode,
-    ...(configResult.config.state.root ? { root: configResult.config.state.root } : {}),
-    metadataRepoRoot: repo.gitRoot,
-  });
-  const snapshot = await loadRunSnapshot(state, runId);
-  return { repoRoot: repo.projectRoot, state, snapshot, config: configResult.config };
-}
-
-function parseAuditEvent(line: string): Event | null {
-  try {
-    return JSON.parse(line) as Event;
-  } catch {
-    return null;
-  }
-}
-
-function formatAuditEvent(event: Event): string {
-  const base = `${event.ts} [${event.level}] ${event.source} ${event.type}`;
-  const message =
-    event.message ?? (event.type === 'log.message' ? event.payload.message : undefined);
-  if (message) {
-    return `${base} - ${message}`;
-  }
-  if (event.error?.message) {
-    return `${base} - ${event.error.message}`;
-  }
-  if (event.type === 'run.step') {
-    return `${base} - ${event.payload.status}: ${event.payload.title}`;
-  }
-  if (event.type === 'run.finished') {
-    return `${base} - ${event.payload.status} (${event.payload.durationMs}ms)`;
-  }
-  return base;
-}
-
-function findLastSuccessfulStep(
-  steps: Record<string, { status?: string; endedAt?: string }>,
-) {
-  return Object.entries(steps)
-    .filter(([, step]) => step?.status === 'done')
-    .sort((a, b) => (b[1]?.endedAt ?? '').localeCompare(a[1]?.endedAt ?? ''))[0]?.[0];
-}
-
-cli
-  .command('agent plan', 'Generate plan')
-  .option('--task <task>', 'Task reference (Linear ID, gh-<number>, or URL)')
-  .action((options: CliOptions) =>
-    withCliContext(options, options.json ? 'json' : 'headless', async (ctx) =>
-      withAgentSessions(Boolean(ctx.config.ai.sessions.persist), (sessions) =>
-        runPlanner(ctx, {
-          ...(options.task ? { taskRef: options.task } : {}),
-          sessions,
-        }),
-      ),
-    ),
-  );
-
-cli
-  .command('agent clarify', 'Answer plan questions')
-  .option('--answer <pair>', 'Answer question (id=value)', { default: [] })
-  .action((options: CliOptions & { answer?: string | string[] }) =>
-    withCliContext(options, options.json ? 'json' : 'headless', async (ctx) =>
-      withAgentSessions(Boolean(ctx.config.ai.sessions.persist), async (sessions) => {
-        const logger = createCliLogger(ctx);
-        const state = await ctx.state.readRunState(ctx.runId);
-        const data = (state?.data as Record<string, unknown>) ?? {};
-        const plan = data['plan'];
-        if (!plan || typeof plan !== 'object') {
-          throw new Error('No plan found in run state. Run agent plan first.');
-        }
-
-        const questions = Array.isArray((plan as { questions?: unknown }).questions)
-          ? ((
-              plan as {
-                questions?: Array<{ id: string; text: string; required?: boolean }>;
-              }
-            ).questions ?? [])
-          : [];
-
-        if (questions.length === 0) {
-          await logger.info('No clarifications needed.');
-          return;
-        }
-
-        const provided = parseAnswerPairs(options.answer);
-        const clarifications = await collectClarifications({
-          questions,
-          answers: {
-            ...(typeof data['clarifications'] === 'object' && data['clarifications']
-              ? (data['clarifications'] as Record<string, string>)
-              : {}),
-            ...provided,
-          },
-        });
-
-        const missingRequired = questions.filter(
-          (question) => question.required !== false && !clarifications[question.id],
-        );
-        if (missingRequired.length > 0) {
-          const ids = missingRequired.map((question) => question.id).join(', ');
-          throw new Error(`Missing required clarifications: ${ids}`);
-        }
-
-        await persistRunState(ctx, ctx.events.mode, (data) => ({
-          ...data,
-          clarifications,
-        }));
-
-        const task = data['task'];
-        const taskRef =
-          typeof data['taskRef'] === 'object' && data['taskRef']
-            ? (data['taskRef'] as { raw?: string }).raw
-            : undefined;
-        const taskId =
-          typeof task === 'object' && task && 'id' in task
-            ? (task as { id?: string }).id
-            : undefined;
-
-        await runPlanner(ctx, {
-          ...(taskRef ? { taskRef } : taskId ? { taskRef: taskId } : {}),
-          clarifications,
-          sessions,
-        });
-      }),
-    ),
-  );
-
-cli
-  .command('agent run', 'Execute the implementation plan')
-  .option('--dry-run', 'Read-only mode: no file changes (default)')
-  .option('--apply', 'Allow file modifications (edits, writes, deletes)')
-  .option('--dangerous', 'Allow shell commands and network access (requires --apply)')
-  .action((options: CliOptions) =>
-    withCliContext(options, options.json ? 'json' : 'headless', async (ctx) =>
-      withAgentSessions(Boolean(ctx.config.ai.sessions.persist), async (sessions) => {
-        const runOptions = {
-          ...(options.dryRun ? { dryRun: true } : {}),
-          ...(options.apply ? { apply: true } : {}),
-          ...(options.dangerous ? { dangerous: true } : {}),
-          sessions,
-        };
-        const shouldReview = await runImplementation(ctx, runOptions);
-        if (shouldReview) {
-          await runReviewLoop(ctx, runOptions);
-        }
-        await runLearningNotes(ctx, { allowApply: Boolean(options.apply) });
-      }),
-    ),
-  );
-
-cli
-  .command('agent resume', 'Resume agent')
-  .action((options: CliOptions) =>
-    withCliContext(options, options.json ? 'json' : 'headless', async (ctx) =>
-      withAgentSessions(Boolean(ctx.config.ai.sessions.persist), () => runRecovery(ctx)),
-    ),
-  );
+registerAgentCommands(cli, {
+  withCliContext,
+  withAgentSessions,
+  persistRunState,
+});
 
 // Shell completion
-cli
-  .command('completion <shell>', 'Generate shell completion script (bash, zsh, fish)')
-  .action(async (shell: string, options: CliOptions) => {
-    const normalized = shell.toLowerCase();
-    const command = 'completion';
-    const instructions: Record<string, string> = {
-      bash: '# Add to ~/.bashrc: eval "$(silvan completion bash)"',
-      zsh: '# Add to ~/.zshrc: eval "$(silvan completion zsh)"',
-      fish: '# Save to ~/.config/fish/completions/silvan.fish',
-    };
-
-    let script: string;
-    switch (shell.toLowerCase()) {
-      case 'bash':
-        script = generateBashCompletion();
-        break;
-      case 'zsh':
-        script = generateZshCompletion();
-        break;
-      case 'fish':
-        script = generateFishCompletion();
-        break;
-      default:
-        throw new Error(`Unknown shell: ${shell}. Supported: bash, zsh, fish`);
-    }
-
-    if (options.json) {
-      await emitJsonSuccess({
-        command,
-        data: {
-          shell: normalized,
-          script,
-          instructions: instructions[normalized] ?? '',
-        },
-      });
-      return;
-    }
-
-    if (options.quiet) {
-      return;
-    }
-
-    console.log(script);
-    const hint = instructions[normalized];
-    if (hint) {
-      console.log(hint);
-    }
-  });
+registerCompletionCommand(cli);
 
 registerConfigCommands(cli, {
   buildConfigOverrides,
@@ -3649,124 +1354,9 @@ registerDoctorCommands(cli, {
   withCliContext,
 });
 
-cli
-  .command('convo show <runId>', 'Show conversation context')
-  .option('--limit <limit>', 'Number of messages to show', { default: '20' })
-  .action((runId: string, options: CliOptions & { limit?: string }) =>
-    withCliContext(options, options.json ? 'json' : 'headless', async (ctx) => {
-      const snapshot = await loadConversationSnapshot(ctx.state, runId);
-      if (!snapshot) {
-        throw new Error(`No conversation found for run ${runId}`);
-      }
-      const limit = Math.max(1, Number(options.limit ?? 20) || 20);
-      const summary = summarizeConversationSnapshot(snapshot, { limit });
-      if (options.json) {
-        await emitJsonSuccess({
-          command: 'convo show',
-          data: { runId, summary },
-          bus: ctx.events.bus,
-          runId: ctx.runId,
-          repoRoot: ctx.repo.repoRoot,
-        });
-        return;
-      }
-      if (options.quiet) {
-        return;
-      }
-      console.log(renderConversationSummary(summary));
-    }),
-  );
-
-cli
-  .command('convo export <runId>', 'Export conversation snapshot')
-  .option('--format <format>', 'json or md', { default: 'json' })
-  .action((runId: string, options: CliOptions & { format?: string }) =>
-    withCliContext(options, options.json ? 'json' : 'headless', async (ctx) => {
-      const snapshot = await loadConversationSnapshot(ctx.state, runId);
-      if (!snapshot) {
-        throw new Error(`No conversation found for run ${runId}`);
-      }
-      const format = (options.format ?? 'json') as 'json' | 'md';
-      if (format !== 'json' && format !== 'md') {
-        throw new Error('Format must be json or md');
-      }
-      const content = exportConversationSnapshot(snapshot, { format });
-      if (options.json) {
-        await emitJsonSuccess({
-          command: 'convo export',
-          data: { runId, format, content },
-          bus: ctx.events.bus,
-          runId: ctx.runId,
-          repoRoot: ctx.repo.repoRoot,
-        });
-        return;
-      }
-      if (options.quiet) {
-        return;
-      }
-      console.log(content);
-    }),
-  );
-
-cli
-  .command('convo optimize <runId>', 'Optimize conversation context')
-  .option('--force', 'Force optimization even if below thresholds')
-  .action((runId: string, options: CliOptions & { force?: boolean }) =>
-    withCliContext(options, options.json ? 'json' : 'headless', async (ctx) => {
-      const store = createConversationStore({
-        runId,
-        state: ctx.state,
-        config: ctx.config,
-        bus: ctx.events.bus,
-        context: buildEmitContext(ctx),
-      });
-      const result = await store.optimize(options.force ? { force: true } : undefined);
-
-      if (options.json) {
-        await emitJsonSuccess({
-          command: 'convo optimize',
-          data: {
-            runId,
-            metrics: result.metrics,
-            backupPath: result.backupPath ?? null,
-          },
-          bus: ctx.events.bus,
-          runId: ctx.runId,
-          repoRoot: ctx.repo.repoRoot,
-        });
-        return;
-      }
-
-      if (options.quiet) {
-        return;
-      }
-
-      const metricRows: Array<[string, string]> = [
-        [
-          'Messages',
-          `${result.metrics.beforeMessages} → ${result.metrics.afterMessages}`,
-        ],
-        ['Tokens', `${result.metrics.beforeTokens} → ${result.metrics.afterTokens}`],
-        ['Saved', String(result.metrics.tokensSaved)],
-        ['Compression', `${Math.round(result.metrics.compressionRatio * 100)}%`],
-        ['Summary', result.metrics.summaryAdded ? 'Added' : 'None'],
-      ];
-      if (result.backupPath) {
-        metricRows.push(['Backup', result.backupPath]);
-      }
-
-      const lines: string[] = [];
-      lines.push(renderSectionHeader('Conversation optimized', { width: 60 }));
-      lines.push(...formatKeyValues(metricRows, { labelWidth: 14 }));
-      lines.push(
-        renderNextSteps([
-          `silvan convo show ${runId}`,
-          `silvan convo export ${runId} --format md`,
-        ]),
-      );
-      console.log(lines.join('\n'));
-    }),
-  );
+registerConversationCommands(cli, {
+  withCliContext,
+});
 
 // Command aliases: short names for common commands
 const COMMAND_ALIASES: Record<string, string> = {
@@ -3880,6 +1470,7 @@ function combineMultiWordCommand(argv: string[]): string[] {
   ];
 }
 
+// eslint-disable-next-line complexity
 export async function run(argv: string[]): Promise<void> {
   const jsonEnabled = argv.includes('--json');
   if (jsonEnabled) {
@@ -4061,3 +1652,4 @@ async function withAgentSessions<T>(
     sessions.close();
   }
 }
+/* eslint-enable max-lines */
